backend/requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.0.0
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
flask-cors
gunicorn

-e .